requirements.txt
# requirements for pipeline_1.py (Project Aegis)
# Generated by GitHub Copilot. Review and pin versions as needed.
# Note: A few imports in the code (e.g. `crewai`, `crewai_tools`, `langchain_google_genai`) may be internal, renamed, or not published on PyPI.
# If a package can't be found on PyPI, install it from the project's private index or source distribution.

# Core HTTP / parsing
requests
feedparser

# Data + ML
numpy
sentence-transformers
transformers

# Vector DB
pinecone

# Databases
pymongo

# News APIs
newsapi-python

# LangChain & LLM integrations
langchain  

# CrewAI agent framework (verify exact package name / install source)
crewai
crewai-tools # imported in code as `crewai_tools` / `crewai.tools` - confirm package name

# Google API auth (used by some Google integrations)
google-api-python-client
google-auth

# Optional helpers
python-dotenv

# Pin exact versions once CI/virtualenv smoke tests pass
